requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
attrs==25.1.0
babel==2.17.0
bel-resources==0.0.3
biopython==1.85
bioregistry==0.12.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
colorama==0.4.6
curies==0.10.6
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
et_xmlfile==2.0.0
greenlet==3.1.1
humanize==4.12.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mordred==1.2.0
more-click==0.1.2
more-itertools==10.6.0
multisplitby==0.0.1
networkx==2.8.8
numpy==1.26.4
numpydoc==1.8.0
oddt==0.7
openbabel-wheel==3.1.1.21
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pystow==0.7.0
python-dateutil==2.9.0.post0
PyTrie==0.4.0
pytz==2025.1
ratelimit==2.2.1
rdkit==2024.9.5
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
roman-numerals-py==3.1.0
rpds-py==0.23.1
scikit-learn==1.4.0
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
